+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.11.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Add `ChangeLogged` plugin
 
 [0.11.0](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.11.0) - 2023-01-31
 ------------------------------------------------------------------------